+++ packages/in_app_purchase/example/ios/in_app_purchase_pluginTests/TranslatorTest.m
@@ -16,6 +16,10 @@
 @property(strong, nonatomic) NSDictionary *paymentMap;
 @property(strong, nonatomic) NSDictionary *transactionMap;
 @property(strong, nonatomic) NSDictionary *errorMap;
+<<<<<<< HEAD
+=======
+@property(strong, nonatomic) NSDictionary *localeMap;
+>>>>>>> master
 
 @end
 
@@ -25,14 +29,14 @@
   self.periodMap = @{@"numberOfUnits" : @(0), @"unit" : @(0)};
   self.discountMap = @{
     @"price" : @1.0,
-    @"currencyCode" : @"USD",
+    @"priceLocale" : [NSLocale.systemLocale toMap],
     @"numberOfPeriods" : @1,
     @"subscriptionPeriod" : self.periodMap,
     @"paymentMode" : @1
   };
   self.productMap = @{
     @"price" : @1.0,
-    @"currencyCode" : @"USD",
+    @"priceLocale" : [NSLocale.systemLocale toMap],
     @"productIdentifier" : @"123",
     @"localizedTitle" : @"title",
     @"localizedDescription" : @"des",
@@ -166,5 +170,13 @@
   NSDictionary *map = [error toMap];
   XCTAssertEqualObjects(map, self.errorMap);
 }
+<<<<<<< HEAD
+=======
+- (void)testLocaleToMap {
+  NSLocale *system = NSLocale.systemLocale;
+  NSDictionary *map = [system toMap];
+  XCTAssertEqualObjects(map[@"currencySymbol"], system.currencySymbol);
+}
+>>>>>>> master
 
 @end