+++ packages/in_app_purchase/ios/Classes/FIAObjectTranslator.h
@@ -61,4 +61,13 @@ NS_ASSUME_NONNULL_BEGIN
 
 @end
 
+<<<<<<< HEAD
+=======
+@interface NSLocale (Coder)
+
+- (nullable NSDictionary *)toMap;
+
+@end
+
+>>>>>>> master
 NS_ASSUME_NONNULL_END