+++ packages/camera/CHANGELOG.md
@@ -1,4 +1,9 @@
+## 0.3.0+1
+
+* Remove initial method call invocation from static camera method.
+
 ## 0.3.0
+
 * **Breaking change**. Migrate from the deprecated original Android Support
   Library to AndroidX. This shouldn't result in any functional changes, but it
   requires any Android apps using this plugin to [also