+++ packages/image_picker/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0+1
+
+* Fix a crash when user calls the plugin in quick succession on Android.
+
 ## 0.5.0
 
 * **Breaking change**. Migrate from the deprecated original Android Support